requirements.txt
# Core production dependencies for Darkstar
pandas>=2.0.0
pyyaml>=6.0
nordpool>=0.2
pytz>=2023.3
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-socketio>=5.11.0
pydantic>=2.0.0
requests>=2.0.0
open-meteo-solar-forecast
pymysql>=1.1.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
astral>=3.2
websockets>=11.0
pulp>=2.9.0
ruamel.yaml>=0.17.0
